// contracts/escrow/typechain-types/AgentForgeEscrow.ts
export declare namespace AgentForgeEscrow {
  export type MilestoneStruct = {
    description: string;
    amount: BigNumberish;
    receiver: AddressLike;
    status: BigNumberish;
    evidence: string;
    agentAddress: AddressLike;
  };

  export type MilestoneStructOutput = [
    description: string,
    amount: bigint,
    receiver: string,
    status: bigint,
    evidence: string,
    agentAddress: string
  ] & {
    description: string;
    amount: bigint;
    receiver: string;
    status: bigint;
    evidence: string;
    agentAddress: string;
  };

  export type EscrowContractStruct = {
    title: string;
    description: string;
    platform: AddressLike;
    serviceProvider: AddressLike;
    approver: AddressLike;
    disputeResolver: AddressLike;
    releaseSigner: AddressLike;
    status: BigNumberish;
    totalFunded: BigNumberish;
    totalReleased: BigNumberish;
    milestones: AgentForgeEscrow.MilestoneStruct[];
  };

  export type EscrowContractStructOutput = [
    title: string,
    description: string,
    platform: string,
    serviceProvider: string,
    approver: string,
    disputeResolver: string,
    releaseSigner: string,
    status: bigint,
    totalFunded: bigint,
    totalReleased: bigint,
    milestones: AgentForgeEscrow.MilestoneStructOutput[]
  ] & {
    title: string;
    description: string;
    platform: string;
    serviceProvider: string;
    approver: string;
    disputeResolver: string;
    releaseSigner: string;
    status: bigint;
    totalFunded: bigint;
    totalReleased: bigint;
    milestones: AgentForgeEscrow.MilestoneStructOutput[];
  };
}